Luotiin Db_queries luokka ja siirrettiiin lokin päivitys siihen.
Luotiin Db_queries luokka ja siirrettiiin lokin päivitys siihen.
