+++ R3_VillagePeople_Mahtimokit/Common.cs
@@ -23,21 +23,6 @@ namespace R3_VillagePeople_Mahtimokit
 
     public class Common_methods
     {
-
-        public void Update_log(string lisatieto_loki)
-        {
-            frm_Main_Window main_window = new frm_Main_Window();
-            // Loki taulun päivitys
-            SqlCommand database_query_loki = new SqlCommand("INSERT INTO [Loki] ([paivittaja], [lisatieto]) " +
-                "VALUES(@paivittaja, @lisatieto_loki)");
-            database_query_loki.Connection = main_window.database_connection;
-            main_window.database_connection.Open();
-            database_query_loki.Parameters.AddWithValue("@paivittaja", Properties.Settings.Default["user_name"].ToString());
-            database_query_loki.Parameters.AddWithValue("@lisatieto_loki", lisatieto_loki);
-            database_query_loki.ExecuteNonQuery();
-            main_window.database_connection.Close();
-        }
-
         public bool Is_email_valid(string email)
         {
             try
