+++ R3_VillagePeople_Mahtimokit/Db_queries.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Text.RegularExpressions;
+using System.Windows.Forms;
+using System.Data.SqlClient;
+
+namespace R3_VillagePeople_Mahtimokit
+{
+    public partial class Db_queries : frm_Main_Window
+    {
+        public void Update_log(string lisatieto_loki)
+        {
+            frm_Main_Window main_window = new frm_Main_Window();
+            // Loki taulun päivitys
+            SqlCommand database_query_loki = new SqlCommand("INSERT INTO [Loki] ([paivittaja], [lisatieto]) " +
+                "VALUES(@paivittaja, @lisatieto_loki)");
+            database_query_loki.Connection = main_window.database_connection;
+            main_window.database_connection.Open();
+            database_query_loki.Parameters.AddWithValue("@paivittaja", Properties.Settings.Default["user_name"].ToString());
+            database_query_loki.Parameters.AddWithValue("@lisatieto_loki", lisatieto_loki);
+            database_query_loki.ExecuteNonQuery();
+            main_window.database_connection.Close();
+        }
+    }
+}
